requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.36
tweepy>=4.14.0
atproto>=0.0.54
beautifulsoup4>=4.12.0
playwright>=1.44.0
